произвел рефакторинг BusinessRuleEditorDialog и GroovyEditorDialog, выделил общую часть в GroovyEditorDialogType
