+++ plugins/ru.runa.gpd/src/ru/runa/gpd/extension/decision/GroovyModel.java
@@ -1,0 +1,5 @@
+package ru.runa.gpd.extension.decision;
+
+public interface GroovyModel {
+
+}
